+++ gamedata/member_love.go
@@ -13,19 +13,19 @@ func loadMemberLoveLevel(gamedata *Gamedata, masterdata_db, serverdata_db *xorm.
 	fmt.Println("Loading MemberLoveLevel")
 	err := masterdata_db.Table("m_member_love_level").OrderBy("love_level").Cols("love_point").Find(&gamedata.MemberLoveLevelLovePoint)
 	utils.CheckErr(err)
-	gamedata.MemberLoveLevelCount = len(gamedata.MemberLoveLevelLovePoint)
-	gamedata.MemberLoveLevelLovePoint = append([]int{0}, gamedata.MemberLoveLevelLovePoint...)
+	gamedata.MemberLoveLevelCount = int32(len(gamedata.MemberLoveLevelLovePoint))
+	gamedata.MemberLoveLevelLovePoint = append([]int32{0}, gamedata.MemberLoveLevelLovePoint...)
 }
 
 func init() {
 	addLoadFunc(loadMemberLoveLevel)
 }
 
-func (gamedata *Gamedata) LoveLevelFromLovePoint(lovePoint int) int {
-	low := 1
+func (gamedata *Gamedata) LoveLevelFromLovePoint(lovePoint int32) int32 {
+	low := int32(1)
 	high := gamedata.MemberLoveLevelCount
-	mid := 0
-	res := 0
+	mid := int32(0)
+	res := int32(0)
 	for low <= high {
 		mid = (low + high) / 2
 		if gamedata.MemberLoveLevelLovePoint[mid] <= lovePoint {
